moved Persistence folders into necessary layer
