+++ src/GameStreamer.Domain/Repositories/IGameStreamRepository.cs
@@ -1,0 +1,30 @@
+﻿namespace GameStreamer.Domain.Repositories;
+
+public interface IGameStreamRepository : IDisposable
+{
+
+    #region Rooms
+
+    //RoomDto? GetFirstIncompleteRoom();
+
+    //RoomDto? GetRoomBy(Guid roomGuid);
+
+    //void AddRoom(RoomDto roomForAdd);
+
+    //void UpdateRoom(RoomDto roomForUpdate);
+
+    //void DeleteRoom(RoomDto roomForDelete);
+
+    #endregion
+
+    #region Players
+
+    //PlayerDto GetPlayerBy(Guid playerDataHashGuid);
+
+    //PlayerDataResponseDTO UpdatePlayer(PlayerDto playerForUpdate);
+
+    //PlayerDataResponseDTO DeletePlayer(PlayerDto playerForDelete);
+
+    #endregion
+
+}
