+++ src/GameStreamer.Infrastructure/Storage_DropLater/GameStreamerDbase/Entities/PlayerEntity.cs
